requirements.txt
typer[all]>=0.5.0,<0.7.0
pandas_plink>=2.2.9,<2.3.0
zarr>=2.11.0,<2.12.0
torch>=1.10.2,<1.13.0
update>=0.0.1,<0.1.0
pytorch-lightning>=1.7.7,<1.8.0
wandb>=0.12.21,<0.14.0
seaborn>=0.11.2,<0.12.0
setuptools==59.5.0
wasabi>=0.10.1,<0.11.0
hydra-core>=1.2.0,<1.3.0
ndjson>=0.3.1,<0.4.0